Added fixes to CouponAPI

+++ TheatreProject.CouponAPI/Models/DTOs/CartHeaderDto.cs
@@ -5,7 +5,7 @@ public class CartHeaderDto
     public Guid Id { get; set; }
     public string? UserId { get; set; }
     public string? CouponCode { get; set; }
-    public double DiscountTotal { get; set; }
-    public double GrandTotal { get; set; }
+    public decimal DiscountTotal { get; set; }
+    public decimal GrandTotal { get; set; }
     public string? Email { get; set; }
 }
